moved profileInfo information to profileSetUp.js in controller directory, still need to do worksamples, basic, and instruments
